removed restlet-based implementation in favour of the fully functional cadc-rest based code provided same SimpleJobManager and RequestPathJobManager for path-specific job persistence and execution
