+++ cadc-uws-server/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java'
     id 'maven'
     id 'maven-publish'
-    id 'com.jfrog.bintray' version '1.8.0'
+    id 'com.jfrog.bintray' version '1.8.4'
 }
 
 repositories {
@@ -24,22 +24,18 @@ test {
 }
 
 dependencies {
-    //compile 'commons-fileupload:commons-fileupload:1.3.+'
     compile 'log4j:log4j:1.2.+'
     compile 'javax.servlet:javax.servlet-api:3.1.+'
     compile 'org.jdom:jdom2:2.+'
-    //compile 'org.restlet.jee:org.restlet:2.0.2'
-    //compile 'org.restlet.jee:org.restlet.ext.fileupload:2.0.2'
     compile 'org.springframework:spring-jdbc:2.5.6.SEC01'
 
     compile 'org.opencadc:cadc-log:[1.0.6,)'
-    compile 'org.opencadc:cadc-util:[1.1,)'
-    compile 'org.opencadc:cadc-vosi:1.+'
-    compile 'org.opencadc:cadc-uws:[1.0.2,)'
+    compile 'org.opencadc:cadc-util:[1.2.8,)'
+    compile 'org.opencadc:cadc-vosi:[1.2,)'
+    compile 'org.opencadc:cadc-uws:[1.0,)'
     compile 'org.opencadc:cadc-rest:[1.2.3,)'
 
     testCompile 'junit:junit:4.+'
-    testCompile 'org.easymock:easymock:3.+'
 
     // Used for DAO tests.
     intTestRuntime 'org.postgresql:postgresql:9.4.1209.jre7'
