--- cadc-uws-server/src/main/java/ca/nrc/cadc/uws/web/restlet/validators/JobFormValidatorImpl.java
@@ -1,188 +1,0 @@
-/*
-************************************************************************
-*******************  CANADIAN ASTRONOMY DATA CENTRE  *******************
-**************  CENTRE CANADIEN DE DONNÉES ASTRONOMIQUES  **************
-*
-*  (c) 2009.                            (c) 2009.
-*  Government of Canada                 Gouvernement du Canada
-*  National Research Council            Conseil national de recherches
-*  Ottawa, Canada, K1A 0R6              Ottawa, Canada, K1A 0R6
-*  All rights reserved                  Tous droits réservés
-*                                       
-*  NRC disclaims any warranties,        Le CNRC dénie toute garantie
-*  expressed, implied, or               énoncée, implicite ou légale,
-*  statutory, of any kind with          de quelque nature que ce
-*  respect to the software,             soit, concernant le logiciel,
-*  including without limitation         y compris sans restriction
-*  any warranty of merchantability      toute garantie de valeur
-*  or fitness for a particular          marchande ou de pertinence
-*  purpose. NRC shall not be            pour un usage particulier.
-*  liable in any event for any          Le CNRC ne pourra en aucun cas
-*  damages, whether direct or           être tenu responsable de tout
-*  indirect, special or general,        dommage, direct ou indirect,
-*  consequential or incidental,         particulier ou général,
-*  arising from the use of the          accessoire ou fortuit, résultant
-*  software.  Neither the name          de l'utilisation du logiciel. Ni
-*  of the National Research             le nom du Conseil National de
-*  Council of Canada nor the            Recherches du Canada ni les noms
-*  names of its contributors may        de ses  participants ne peuvent
-*  be used to endorse or promote        être utilisés pour approuver ou
-*  products derived from this           promouvoir les produits dérivés
-*  software without specific prior      de ce logiciel sans autorisation
-*  written permission.                  préalable et particulière
-*                                       par écrit.
-*                                       
-*  This file is part of the             Ce fichier fait partie du projet
-*  OpenCADC project.                    OpenCADC.
-*                                       
-*  OpenCADC is free software:           OpenCADC est un logiciel libre ;
-*  you can redistribute it and/or       vous pouvez le redistribuer ou le
-*  modify it under the terms of         modifier suivant les termes de
-*  the GNU Affero General Public        la “GNU Affero General Public
-*  License as published by the          License” telle que publiée
-*  Free Software Foundation,            par la Free Software Foundation
-*  either version 3 of the              : soit la version 3 de cette
-*  License, or (at your option)         licence, soit (à votre gré)
-*  any later version.                   toute version ultérieure.
-*                                       
-*  OpenCADC is distributed in the       OpenCADC est distribué
-*  hope that it will be useful,         dans l’espoir qu’il vous
-*  but WITHOUT ANY WARRANTY;            sera utile, mais SANS AUCUNE
-*  without even the implied             GARANTIE : sans même la garantie
-*  warranty of MERCHANTABILITY          implicite de COMMERCIALISABILITÉ
-*  or FITNESS FOR A PARTICULAR          ni d’ADÉQUATION À UN OBJECTIF
-*  PURPOSE.  See the GNU Affero         PARTICULIER. Consultez la Licence
-*  General Public License for           Générale Publique GNU Affero
-*  more details.                        pour plus de détails.
-*                                       
-*  You should have received             Vous devriez avoir reçu une
-*  a copy of the GNU Affero             copie de la Licence Générale
-*  General Public License along         Publique GNU Affero avec
-*  with OpenCADC.  If not, see          OpenCADC ; si ce n’est
-*  <http://www.gnu.org/licenses/>.      pas le cas, consultez :
-*                                       <http://www.gnu.org/licenses/>.
-*
-*  $Revision: 4 $
-*
-************************************************************************
-*/
-
-
-package ca.nrc.cadc.uws.web.restlet.validators;
-
-import org.restlet.data.Form;
-
-import java.util.Map;
-import java.util.HashMap;
-import java.text.ParseException;
-
-import ca.nrc.cadc.uws.web.validators.FormValidator;
-import ca.nrc.cadc.uws.JobAttribute;
-import ca.nrc.cadc.uws.ExecutionPhase;
-import ca.nrc.cadc.date.DateUtil;
-import ca.nrc.cadc.util.StringUtil;
-
-import java.text.DateFormat;
-
-
-/**
- * Form validator implementation specific to a Job Form.
- */
-public class JobFormValidatorImpl implements FormValidator
-{
-    final Form form;
-
-
-    /**
-     * Default constructor.
-     *
-     * @param form      The form to operate on.
-     */
-    public JobFormValidatorImpl(final Form form)
-    {
-        this.form = form;
-    }
-
-    
-    /**
-     * Validate the given Form.
-     *
-     * @return Mapping of fields to error messages, if necessary, or
-     *         empty Map otherwise, never NULL.
-     */
-    public Map<String, String> validate()
-    {
-        final Map<String, String> errors = new HashMap<String, String>();
-        final Map<String, String> valuesMap = getForm().getValuesMap();
-
-        // Check for a valid Exection Phase.
-        if (valuesMap.containsKey(
-                JobAttribute.EXECUTION_PHASE.getAttributeName().toUpperCase()))
-        {
-            try
-            {
-                ExecutionPhase.valueOf(
-                        valuesMap.get(JobAttribute.EXECUTION_PHASE.
-                                getAttributeName().toUpperCase()).
-                                toUpperCase());
-            }
-            catch (final IllegalArgumentException e)
-            {
-                errors.put(JobAttribute.EXECUTION_PHASE.
-                        getAttributeName().toUpperCase(),
-                           "Execution Phase is mandatory.");
-            }
-        }
-//
-//        if (!valuesMap.containsKey(JobAttribute.EXECUTION_DURATION.
-//                getAttributeName().toUpperCase()))
-//        {
-//            errors.put(JobAttribute.EXECUTION_DURATION.
-//                    getAttributeName().toUpperCase(),
-//                       "Execution Duration is mandatory.");
-//        }
-
-        final String destruction = valuesMap.get(
-                JobAttribute.DESTRUCTION_TIME.getAttributeName().toUpperCase());
-
-        if (StringUtil.hasText(destruction) && !isValidDate(destruction))
-        {
-            errors.put(JobAttribute.DESTRUCTION_TIME.
-                    getAttributeName().toUpperCase(),
-                       "Destruction Time is a mandatory date in the format "
-                       + DateUtil.IVOA_DATE_FORMAT);
-        }
-
-//        final String quote = getForm().getFirstValue(
-//                JobAttribute.QUOTE.getAttributeName().toUpperCase());
-//
-//        if (!StringUtil.hasText(quote) || !isValidDate(quote))
-//        {
-//            errors.put(JobAttribute.QUOTE.getAttributeName().toUpperCase(),
-//                       "Quote Time is a mandatory date in the format "
-//                       + DateUtil.ISO8601_DATE_FORMAT);
-//        }
-
-        return errors;
-    }
-
-    protected boolean isValidDate(final String dateString)
-    {
-        try
-        {
-            DateFormat df = DateUtil.getDateFormat(DateUtil.IVOA_DATE_FORMAT, DateUtil.UTC);
-            df.parse(dateString);
-        }
-        catch (ParseException e)
-        {
-            return false;
-        }
-        return true;
-    }
-
-
-    public Form getForm()
-    {
-        return form;
-    }
-}
