+++ example-uws/build.gradle
@@ -27,19 +27,12 @@ war {
 }
 
 dependencies {
-    compile 'commons-fileupload:commons-fileupload:1.3.+'
     compile 'log4j:log4j:1.2.+'
-    compile 'javax.servlet:javax.servlet-api:3.1.+'
-    compile 'org.jdom:jdom2:2.+'
-    compile 'org.restlet.jee:org.restlet:2.0.2'
-    compile 'org.restlet.jee:org.restlet.ext.fileupload:2.0.2'
-    compile 'org.springframework:spring-beans:2.5.6.SEC01'
-    compile 'xerces:xercesImpl:2.+'
 
     compile 'org.opencadc:cadc-log:1.+'
     compile 'org.opencadc:cadc-util:1.+'
     compile 'org.opencadc:cadc-uws:1.+'
-    compile 'org.opencadc:cadc-uws-server:[1.1,)'
+    compile 'org.opencadc:cadc-uws-server:[1.2,)'
     
     testCompile 'junit:junit:[4.0,5.0)'
     intTestCompile 'org.opencadc:cadc-test-uws:[1.1,)'
